Rename segment_traverse
