+++ scanner/marking.h
