+++ testing_system/test.c
@@ -1,21 +1,19 @@
-#include "../scanner/segment_traverse.h"
+#include "../scanner/marking.h"
 #include "../allocator/allocator.h"
 #include <stdio.h>
 
 typedef struct A { int* a1; int* a2; int* a3; int* a4; }A;
 
 int main() {
-	A** a = allocate_new_object(32);
-	for (int i = 0; i < 8; i++) {
-		if (i % 2 == 0) {
+	A* a[8 * 16];
+	for (int i = 0; i < 8 * 16; i++) {
+		if (i % 2 == 1) {
 			allocate_new_object(32);
 			continue;
 		}
 		a[i] = (A*)allocate_new_object(32);
 	}
 	segment_traverse(end_rsp_value, start_rsp_value);
-	printf("size = %d\n", sizeof(a));
-	show_bitmap((size_t)a);
+	show_bitmap((size_t*)*a);
 	return 0;
-
 }
